[FIX] collection('db/path') now checks startsWith instead of contains. [FIX] collection('db/path') adds a preceeding slash to the document name for comparison, so files situated one level down the folder hierarchy are correctly added to the context.
Testcase PathTest#withoutIndexTestWeek has been modified to reflect the new behavior.
This fix is open for discussion, the idea of using startsWith just felt more natural to me. Anyway for documents in the first hierarchy level of a collection the second [FIX] should persist, 
even if s/o decides to revert the startsWith ;-)