Made Access to MetaData thread-safe
MetaData includes an instance of Prop, which in turn had to be made
thread-safe (by making AProp thread-safe). All of this is necessary
because access to the same MetaData instance takes place from many
different threads in an unsynchronized manner.
On most variables, we are simply using volatile, which is not strictly
correct in all cases because some variables depend on each other. We
might later want to review this, but the volatile version is in any case
much safer than what we currently have.